feat: replace csv generator by bundle generator (#247)
* feat: replace csv generator by bundle generator

Fix https://github.com/quarkiverse/quarkus-operator-sdk/issues/243

* fix build in dev mode

* fix path of dockerbuild

* fix annotations file

* Change bundle default configuration
