+++ bom/pom.xml
@@ -76,12 +76,12 @@
         </dependency>
         <dependency>
           <groupId>io.quarkiverse.operatorsdk</groupId>
-          <artifactId>quarkus-operator-sdk-csv-generator</artifactId>
+          <artifactId>quarkus-operator-sdk-bundle-generator</artifactId>
           <version>${project.version}</version>
         </dependency>
         <dependency>
           <groupId>io.quarkiverse.operatorsdk</groupId>
-          <artifactId>quarkus-operator-sdk-csv-generator-deployment</artifactId>
+          <artifactId>quarkus-operator-sdk-bundle-generator-deployment</artifactId>
           <version>${project.version}</version>
         </dependency>
         <dependency>
