+++ bundle-generator/deployment/src/main/java/io/quarkiverse/operatorsdk/bundle/deployment/GeneratedBundleBuildItem.java
@@ -1,0 +1,6 @@
+package io.quarkiverse.operatorsdk.bundle.deployment;
+
+import io.quarkus.builder.item.SimpleBuildItem;
+
+public final class GeneratedBundleBuildItem extends SimpleBuildItem {
+}
