+++ bundle-generator/runtime/src/main/java/io/quarkiverse/operatorsdk/bundle/runtime/SharedCSVMetadata.java
@@ -1,0 +1,4 @@
+package io.quarkiverse.operatorsdk.bundle.runtime;
+
+public interface SharedCSVMetadata {
+}
