--- csv-generator/deployment/src/main/java/io/quarkiverse/operatorsdk/csv/deployment/GeneratedCSVBuildItem.java
@@ -1,6 +1,0 @@
-package io.quarkiverse.operatorsdk.csv.deployment;
-
-import io.quarkus.builder.item.SimpleBuildItem;
-
-public final class GeneratedCSVBuildItem extends SimpleBuildItem {
-}
