--- csv-generator/deployment/src/main/java/io/quarkiverse/operatorsdk/csv/deployment/builders/ManifestsBuilder.java
@@ -1,68 +1,0 @@
-package io.quarkiverse.operatorsdk.csv.deployment.builders;
-
-import java.io.IOException;
-import java.util.List;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
-import com.fasterxml.jackson.dataformat.yaml.YAMLGenerator;
-
-import io.fabric8.kubernetes.api.model.ServiceAccount;
-import io.fabric8.kubernetes.api.model.apps.Deployment;
-import io.fabric8.kubernetes.api.model.rbac.ClusterRole;
-import io.fabric8.kubernetes.api.model.rbac.ClusterRoleBinding;
-import io.fabric8.kubernetes.api.model.rbac.Role;
-import io.fabric8.kubernetes.api.model.rbac.RoleBinding;
-import io.quarkiverse.operatorsdk.csv.deployment.AugmentedResourceInfo;
-
-public abstract class ManifestsBuilder {
-
-    protected static final ObjectMapper YAML_MAPPER;
-
-    static {
-        YAML_MAPPER = new ObjectMapper((new YAMLFactory()).enable(YAMLGenerator.Feature.MINIMIZE_QUOTES)
-                .enable(YAMLGenerator.Feature.ALWAYS_QUOTE_NUMBERS_AS_STRINGS)
-                .disable(YAMLGenerator.Feature.WRITE_DOC_START_MARKER));
-        YAML_MAPPER.configure(SerializationFeature.INDENT_OUTPUT, true);
-        YAML_MAPPER.configure(SerializationFeature.WRITE_NULL_MAP_VALUES, false);
-        YAML_MAPPER.configure(SerializationFeature.WRITE_EMPTY_JSON_ARRAYS, false);
-    }
-
-    private final String controllerName;
-    private final String csvGroupName;
-
-    public ManifestsBuilder(AugmentedResourceInfo cri) {
-        controllerName = cri.getControllerName();
-        csvGroupName = cri.getCsvGroupName();
-    }
-
-    public abstract String getFileName();
-
-    public abstract byte[] getYAMLData(List<ServiceAccount> serviceAccounts, List<ClusterRoleBinding> clusterRoleBindings,
-            List<ClusterRole> clusterRoles, List<RoleBinding> roleBindings, List<Role> roles,
-            List<Deployment> deployments) throws IOException;
-
-    public String getControllerName() {
-        return controllerName;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-
-        ManifestsBuilder that = (ManifestsBuilder) o;
-
-        return getFileName().equals(that.getFileName());
-    }
-
-    @Override
-    public int hashCode() {
-        return getFileName().hashCode();
-    }
-}
