--- csv-generator/runtime/src/main/java/io/quarkiverse/operatorsdk/csv/runtime/SharedCSVMetadata.java
@@ -1,4 +1,0 @@
-package io.quarkiverse.operatorsdk.csv.runtime;
-
-public interface SharedCSVMetadata {
-}
