--- csv-generator/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,13 +1,0 @@
-name: Quarkus - Operator SDK - CSV Generator
-description: CSV Generation for Java Operator SDK operators
-metadata:
-keywords:
-  - operator
-  - kubernetes
-  - openshift
-  - olm
-  - csv
-#  guide: ...
-categories:
-  - cloud
-#  status: "preview"
