+++ integration-tests/pom.xml
@@ -24,7 +24,7 @@
     </dependency>
     <dependency>
       <groupId>io.quarkiverse.operatorsdk</groupId>
-      <artifactId>quarkus-operator-sdk-csv-generator</artifactId>
+      <artifactId>quarkus-operator-sdk-bundle-generator</artifactId>
     </dependency>
      <dependency>
        <groupId>io.quarkus</groupId>
