+++ pom.xml
@@ -27,7 +27,7 @@
         <module>common</module>
         <module>common-deployment</module>
         <module>core</module>
-        <module>csv-generator</module>
+        <module>bundle-generator</module>
       </modules>
     </profile>
     <profile>
