+++ samples/joke/README.md
@@ -17,12 +17,12 @@ cluster. You can just follow the steps below to get started quickly:
 - Run `mvn -Dquickly` on the root directory to build the project locally and move to this sample directory.
 - Run `mvn package` to build the operator. This will automatically generate several resources for
   you, in particular the CRDs associated with the custom resources we will be dealing with. These
-  CRDs are generated in `target/manifests/` and come in `v1` version which correspond to the versions of the CRD spec. Note that older clusters might need the version `v1beta1` which is no longer generated by default. To generate it, you need to use the property `quarkus.operator-sdk.crd.versions=v1beta1,v1`, so both versions `v1` and `v1beta1` are generated.
+  CRDs are generated in `target/kubernetes/` and come in `v1` version which correspond to the versions of the CRD spec. Note that older clusters might need the version `v1beta1` which is no longer generated by default. To generate it, you need to use the property `quarkus.operator-sdk.crd.versions=v1beta1,v1`, so both versions `v1` and `v1beta1` are generated.
 - Deploy the CRDs to your cluster (requires cluster admin privileges, while this shouldn't be an
   issue for most "testing" clusters such as `minikube` or `kind`, you might need to log in to your
   OpenShift clusters with an admin account):
   ```sh
-  kubectl apply -f target/manifests/jokerequests.samples.javaoperatorsdk.io-v1.yml
+  kubectl apply -f target/kubernetes/jokerequests.samples.javaoperatorsdk.io-v1.yml
   kubectl apply -f src/main/k8s/jokes.samples.javaoperatorsdk.io-v1.yml
   ```           
 - Launch the app in dev mode: `mvn quarkus:dev`
@@ -51,7 +51,7 @@ Make sure you have installed the [opm](https://github.com/operator-framework/ope
 2. Generate the Operator image and bundle manifests
 
 This example uses the [Quarkus Jib container image](https://quarkus.io/guides/container-image#jib) extension to build the Operator image. 
-Also, the Quarkus Operator SDK provides the `quarkus-operator-sdk-csv-generator` extension that generates the Operator bundle manifests at `target/manifests`.
+Also, the Quarkus Operator SDK provides the `quarkus-operator-sdk-bundle-generator` extension that generates the Operator bundle manifests at `target/bundle`.
 So, you simply need to run the next Maven command to build and push the operator image, and also to generate the bundle manifests:
 
 ```shell
@@ -59,8 +59,17 @@ mvn clean package -Dquarkus.container-image.build=true \
     -Dquarkus.container-image.push=true \
     -Dquarkus.container-image.registry=<your container registry. Example: quay.io> \
     -Dquarkus.container-image.group=<your container registry namespace> \
-    -Dquarkus.kubernetes.namespace=<the kubernetes namespace where you will deploy the operator>
+    -Dquarkus.kubernetes.namespace=<the kubernetes namespace where you will deploy the operator> \
+    -Dquarkus.operator-sdk.bundle.package-name=<the name of the package that bundle image belongs to> \ 
+    -Dquarkus.operator-sdk.bundle.channels=<the list of channels that bundle image belongs to>
 ```
+
+For example, if we want to name the package `joke-operator` and use the `alpha` channels, we would need to append the properties `-Dquarkus.operator-sdk.bundle.package-name=joke-operator -Dquarkus.operator-sdk.bundle.channels=alpha`.
+
+---
+**NOTE**
+Find more information about channels and packages in [this link](https://olm.operatorframework.io/docs/best-practices/channel-naming/#channels).
+---
 
 ---
 **NOTE**
@@ -70,28 +79,11 @@ If you're using an insecure container registry, you'd also need to append the ne
 2. Build the Operator Bundle image
 
 An Operator Bundle is a container image that stores Kubernetes manifests and metadata associated with an operator. You can find more information about this in [here](https://olm.operatorframework.io/docs/tasks/creating-operator-bundle/). 
-In the previous step, we generated the manifests at `target/manifests`, but we still need to generate the `metadata/annotations.yaml` file and the final Operator Bundle image. For doing this, we will use the [opm](https://github.com/operator-framework/operator-registry) utility:
-
-```shell
-opm alpha bundle generate \
-    --directory target/manifests \
-    --package <The name of the package that bundle image belongs to> \ 
-    --channels <The list of channels that bundle image belongs to> \
-    --default <The default channel for the bundle image>
-```
- 
-For example, if we want to name the package `joke` and use the `alpha` channels, the command would look like as `opm alpha bundle generate --directory target/manifests --package joke --channels alpha --default alpha`.
-
----
-**NOTE**
-Find more information about channels and packages in [this link](https://olm.operatorframework.io/docs/best-practices/channel-naming/#channels).
----
-
-The above command will generate a `bundle.Dockerfile` Dockerfile file that you will use to build and push the final Operator Bundle image to your container registry:
+In the previous step, we generated the bundle manifests at `target/bundle` which includes a ready-to-use Dockerfile at `target/bundle/bundle.Dockerfile` file that you will use to build and push the final Operator Bundle image to your container registry:
 
 ```shell
 JOKE_BUNDLE_IMAGE=<your container registry>/<your container registry namespace>/<bundle image name>:<tag>
-docker build -t $JOKE_BUNDLE_IMAGE -f bundle.Dockerfile .
+docker build -t $JOKE_BUNDLE_IMAGE -f target/bundle/bundle.Dockerfile target/bundle
 docker push $JOKE_BUNDLE_IMAGE
 ```
