+++ samples/joke/pom.xml
@@ -22,7 +22,7 @@
     </dependency>
     <dependency>
       <groupId>io.quarkiverse.operatorsdk</groupId>
-      <artifactId>quarkus-operator-sdk-csv-generator</artifactId>
+      <artifactId>quarkus-operator-sdk-bundle-generator</artifactId>
     </dependency>
     <dependency>
       <groupId>io.quarkus</groupId>
@@ -39,7 +39,7 @@
     <!--  This dependency is needed only to ensure proper building order so that this module is build after the CSV extension -->
     <dependency>
       <groupId>io.quarkiverse.operatorsdk</groupId>
-      <artifactId>quarkus-operator-sdk-csv-generator-deployment</artifactId>
+      <artifactId>quarkus-operator-sdk-bundle-generator-deployment</artifactId>
       <scope>provided</scope>
     </dependency>
     <dependency>
