+++ samples/joke/src/main/resources/application.properties
@@ -1,4 +1,5 @@
 joke-api/mp-rest/url=https://v2.jokeapi.dev/joke
 quarkus.operator-sdk.crd.apply=true
-quarkus.operator-sdk.generate-csv=true
 quarkus.container-image.builder=jib
+quarkus.operator-sdk.bundle.channels=alpha
+quarkus.operator-sdk.bundle.package-name=joke-operator
