+++ samples/mysql-schema/pom.xml
@@ -20,7 +20,7 @@
     </dependency>
     <dependency>
       <groupId>io.quarkiverse.operatorsdk</groupId>
-      <artifactId>quarkus-operator-sdk-csv-generator</artifactId>
+      <artifactId>quarkus-operator-sdk-bundle-generator</artifactId>
     </dependency>
     <dependency>
       <groupId>io.quarkus</groupId>
@@ -36,7 +36,7 @@
     </dependency>
     <dependency>
       <groupId>io.quarkiverse.operatorsdk</groupId>
-      <artifactId>quarkus-operator-sdk-csv-generator-deployment</artifactId>
+      <artifactId>quarkus-operator-sdk-bundle-generator-deployment</artifactId>
       <scope>provided</scope>
     </dependency>
     <dependency>
