+++ samples/mysql-schema/src/main/resources/application.properties
@@ -1,5 +1,4 @@
 quarkus.operator-sdk.crd.apply=true
-quarkus.operator-sdk.generate-csv=true
 quarkus.container-image.builder=jib
 quarkus.native.additional-build-args=--initialize-at-run-time=org.apache.commons.lang3.RandomStringUtils
